Separate USBPhyEvents and USBPhyTypes from USBPhy
Break apart USBPhy into three separate header files. Also make some
doxygen updates.
